[#43][feature/view-comments] 댓글 작성, 대댓글 작성 뷰 화면 구성 (#46)
* feat: 댓글 작성 화면 구성 (#43)

* feat: 대댓글 작성 기능 구현 (#43)

* refactor: Comment DB 스키마 변경 및 코드 수정 (#43)

+++ src/main/java/com/spring/blog/comment/presentation/dto/CommentResponse.java
@@ -8,7 +8,7 @@ public class CommentResponse {
     private Long id;
     private String author;
     private String content;
-    private Integer depth;
+    private Long depth;
     private LocalDateTime createdDate;
 
     private CommentResponse() {
@@ -18,7 +18,7 @@ public class CommentResponse {
         Long id,
         String author,
         String content,
-        Integer depth,
+        Long depth,
         LocalDateTime createdDate
     ) {
         this.id = id;
@@ -50,7 +50,7 @@ public class CommentResponse {
         return content;
     }
 
-    public Integer getDepth() {
+    public Long getDepth() {
         return depth;
     }
